+++ locus-api-android-sample/src/main/java/com/asamm/locus/api/sample/ActivityGeocacheTools.java
@@ -3,10 +3,10 @@ package com.asamm.locus.api.sample;
 import android.app.AlertDialog;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v4.app.FragmentActivity;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import androidx.fragment.app.FragmentActivity;
 import locus.api.android.ActionTools;
 import locus.api.android.utils.LocusUtils;
 import locus.api.android.utils.LocusUtils.LocusVersion;