+++ locus-api-android-sample/src/main/java/com/asamm/locus/api/sample/pages/PageWelcomeFragment.java
@@ -1,12 +1,13 @@
 package com.asamm.locus.api.sample.pages;
 
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
 import com.asamm.locus.api.sample.R;
+
+import androidx.fragment.app.Fragment;
 
 /**
  * Created by menion on 29/08/2016.